Splash Screen Revisions
Updated the SplashScreen .Close Function to gracefully kill the thread. To prevent permission issues/access violations with setting it to null the Thread.Abort function is being used first. Additionally I added a try/catch around the .close logic to try not to throw up any exceptions during these attempts, but instead to display a messagebox alerting the user to the issue.

I Also added a try catch in the Initial Load of the application so that any exceptions are handled in a more graceful manner.